backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
cohere==4.3.1
qdrant-client==1.8.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.23.0
httpx==0.25.2
python-frontmatter==1.0.0